+++ ski-tab/src/app/components/settings-competitions/settings-competitions.component.ts
@@ -1,9 +1,10 @@
 import { Component } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { JwtHelperService } from '@auth0/angular-jwt';
 import { CookieService } from 'ngx-cookie-service';
-import { Competitions } from 'src/app/models/competitions';
 import { CompetitionsService } from 'src/app/services/competitions/competitions.service';
-import { UsersService } from 'src/app/services/users/users.service';
+import { MessengerService } from '../../services/messenger/messenger.service';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-settings-competitions',
@@ -14,9 +15,10 @@ export class SettingsCompetitionsComponent {
 
   constructor(
     private competitionService: CompetitionsService,
-    private usersService: UsersService,
     private cookieService: CookieService,
     private jwt: JwtHelperService,
+    private dialog: MatDialog,
+    private messengerService: MessengerService,
   ){}
 
   userToken = this.cookieService.get('accessToken');
@@ -39,6 +41,17 @@ export class SettingsCompetitionsComponent {
   }
 
   deleteCompetition(competitionId:any){
-    this.competitionService.deleteCompetition(competitionId);
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data:{
+        title: 'Delete competition?'
+      }
+    })
+    dialogRef.afterClosed().subscribe( check => {
+      if(check) {
+        this.competitionService.deleteCompetition(competitionId);
+        this.messengerService.showNotification("Competition deleted!", 2000);
+      }    
+    })
+    
   }
 }
